* main:
  Feature: Trade (#10)

# Conflicts:
#	cmd/tools/dump.go
#	internal/native/event_mapper.go
#	internal/native/list_mapper.go
#	pkg/parser/argument_parser.go

+++ pkg/events/candle/candle.go
@@ -182,7 +182,7 @@ func (q *Candle) Time() int64 {
 }
 
 func (q *Candle) SetTime(value int64) {
-	q.index = (int64(timeutil.GetSecondsFromTime(value)) << 32) |
+	q.index = (timeutil.GetSecondsFromTime(value) << 32) |
 		int64(timeutil.GetMillisFromTime(value)<<22) |
 		q.Sequence()
 }
